sys/analog_util: Use int32_t for adc_util_map args

+++ sys/analog_util/adc_util.c
@@ -21,6 +21,7 @@
  */
 
 #include <stdint.h>
+#include <inttypes.h>
 #include "cpu.h"
 #include "periph/adc.h"
 #include "analog_util.h"
@@ -50,16 +51,16 @@ static unsigned int _adc_res_bits(adc_res_t res)
     }
 }
 
-int adc_util_map(int sample, adc_res_t res, int min, int max)
+int32_t adc_util_map(int sample, adc_res_t res, int32_t min, int32_t max)
 {
     /* Using 64 bit signed int as intermediate to prevent overflow when range
      * multiplied by sample requires more bits than int offers */
-    int scaled = (((int64_t)(max - min) * sample) >> _adc_res_bits(res));
-    DEBUG("scaled: %d\n", scaled);
+    int32_t scaled = (((int64_t)(max - min) * sample) >> _adc_res_bits(res));
+    DEBUG("scaled: %" PRId32 "\n", scaled);
     return (min + scaled);
 }
 
 float adc_util_mapf(int sample, adc_res_t res, float min, float max)
 {
-    return ((((max - min) * sample) / (1 << _adc_res_bits(res))) + min);
+    return ((((max - min) * sample) / ((int32_t)1L << _adc_res_bits(res))) + min);
 }
